bug: (python) fix multiple embedding functions bug
Closes #594

The embedding functions are pydantic models so multiple
instances with the same parameters are considered ==, which
means that if you have multiple embedding columns it's possible
for the embeddings to get overwritten. Instead we use `is` instead
of == to avoid this problem.

testing: modified unit test to include this case