Update table.py
If an empty list is passed as an arg to the data param, it will pass the isinstance test on line 53 but will fail with an IndexError: list index out of range on line 55.  Checking to see if data is not empty will solve this error as well as other potential downstream errors.  Throwing a ValueError if the data is empty, but open to other suggestions on how to gracefully handle.