feat(python): add option to flatten output in to_pandas
Closes Lance#1738

We add a `flatten` parameter to the signature of `to_pandas`.
By default this is None and does nothing.
If set to True or -1, then LanceDB will flatten structs before
converting to a pandas dataframe. All nested structs are also flattened.
If set to any positive integer, then LanceDB will flatten
structs up to the specified level of nesting.