feat: introduce ArrowNative wrapper struct for adding data that is already a RecordBatchReader (#1139)
In
https://github.com/lancedb/lancedb/commit/3d30dd2bf05d926444b5ee4b8ba5aa182f337d34
I added a new `IntoArrow` trait for adding data into a table.
Unfortunately, it seems my approach for implementing the trait for
"things that are already record batch readers" was flawed. This PR
corrects that flaw and, conveniently, removes the need to box readers at
all (though it is ok if you do).