feat: expose table trait (#2097)
Similar to
https://github.com/lancedb/lancedb/commit/c269524b2f394a6c8508436840bce17d884a42c0
this PR reworks and exposes an internal trait (this time
`TableInternal`) to be a public trait. These two PRs together should
make it possible for others to integrate LanceDB on top of other
catalogs.

This PR also adds a basic `TableProvider` implementation for tables,
although some work still needs to be done here (pushdown not yet
enabled).