Fix excessive object storage handle creation in create_table
This fixes two bugs with create_table storage handle reuse. First issue
is, the database object did not previously carry a session that
create_table operations could reuse for create_table operations.

Second issue is, the inheritance logic for create_table and open_table
was causing empty storage options (i.e Some({})) to get sent, instead of
None. Lance handles these differently:

* When None is set, the object store held in the session's storage
  registry that was created at "connect" is used. This value stays in
  the cache long-term (probably as long as the db reference is held).
* When Some({}) is sent, LanceDB will create a new connection and cache
  it for an empty key. However, that cached value will remain valid only
  as long as the client holds a reference to the table. After that, the
  cache is poisoned and the next create_table with the same key, will
  create a new connection. This confounds reuse if e.g python gc's the
  table object before another table is created.

My feeling is that the second path, if intentional, is probably meant to
serve cases where tables are overriding settings and the cached
connection is assumed not to be generally applicable. The bug is we were
engaging that mechanism for all tables.